api/requirements.txt
Flask
Flask-Compress
Flask-Cors
Flask-RESTful
marshmallow
psycopg2
uwsgi
sqlalchemy
flask-apispec
Pillow